+++ R/interactive_annotate.R
@@ -16,6 +16,7 @@
 #' @param range_max_kerning define maximum of kerning in slider. must be positive.
 #' @param resolution resolution of slider
 #' @param return_param If return_param is TRUE, returns a list of values of location, degrees, size, weight, and kerning. If return_param is FALSE, returns a magick image object.
+#' @param scale geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.
 #' @return a magick image object or a list of values of location, degrees, size, weight, and kerning
 #' @author Shota Ochi
 #' @export
@@ -23,11 +24,15 @@
 #' \donttest{
 #' interactive_annotate(wizard, "hello")
 #' }
-interactive_annotate <- function(image, text, gravity = "northwest", font = "", style = "normal", decoration = NULL, color = NULL, strokecolor = NULL, boxcolor = NULL, range_max_size = 1000, range_max_weight = 850, range_max_kerning = 300, resolution = 0.1, return_param = FALSE)
+interactive_annotate <- function(image, text, gravity = "northwest", font = "", style = "normal", decoration = NULL, color = NULL, strokecolor = NULL, boxcolor = NULL, range_max_size = 1000, range_max_weight = 850, range_max_kerning = 300, resolution = 0.1, return_param = FALSE, scale)
 {
   # image must be convreted into png to avoid the error of tkimage.create function
   image_original <- image
-  image <- as.list(image)[[1]] %>% image_convert(format = "png")
+  image <- image_convert(as.list(image)[[1]], format = "png")
+  if (!missing(scale))
+  {
+    image <- image_scale(image, scale)
+  }
   
   # make initial output
   inix <- 0