+++ R/interactive_fill.R
@@ -7,6 +7,7 @@
 #' @param refcolor if set, fuzz color distance will be measured against this color, not the color of the starting point. Any color (within fuzz color distance of the given refcolor), connected to starting point will be replaced with the color. If the pixel at the starting point does not itself match the given refcolor (according to fuzz) then no action will be taken.
 #' @param resolution resolution of slider of fuzz
 #' @param return_param If return_param is TRUE, returns a list values of point and fuzz. If return_param is FALSE, returns a magick image object.
+#' @param scale geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.
 #' @return a magick image object or a list of values of point and fuzz
 #' @author Shota Ochi
 #' @export
@@ -14,11 +15,15 @@
 #' \donttest{
 #' interactive_fill(wizard, "black")
 #' }
-interactive_fill <- function(image, color, refcolor = NULL, resolution = 0.1, return_param = FALSE)
+interactive_fill <- function(image, color, refcolor = NULL, resolution = 0.1, return_param = FALSE, scale)
 {
   # image must be convreted into png to avoid the error of tkimage.create function
   image_original <- image
-  image <- as.list(image)[[1]] %>% image_convert(format = "png")
+  image <- image_convert(as.list(image)[[1]], format = "png")
+  if (!missing(scale))
+  {
+    image <- image_scale(image, scale)
+  }
   
   # make initial output
   inix <- 1