+++ R/interactive_implode.R
@@ -6,6 +6,7 @@
 #' @param range_max define maximum in slider. must be positive.
 #' @param resolution resolution of slider
 #' @param return_param If return_param is TRUE, returns value of factor. If return_param is FALSE, returns a magick image object.
+#' @param scale geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.
 #' @return a magick image object or value of factor
 #' @author Shota Ochi
 #' @export
@@ -13,11 +14,15 @@
 #' \donttest{
 #' interactive_implode(wizard)
 #' }
-interactive_implode <- function(image, range_max = 1, resolution = 0.1, return_param = FALSE)
+interactive_implode <- function(image, range_max = 1, resolution = 0.1, return_param = FALSE, scale)
 {
   # image must be convreted into png to avoid the error of tkimage.create function
   image_original <- image
-  image <- as.list(image)[[1]] %>% image_convert(format = "png")
+  image <- image_convert(as.list(image)[[1]], format = "png")
+  if (!missing(scale))
+  {
+    image <- image_scale(image, scale)
+  }
   
   # make initial output
   iniv <- 0