+++ tests/testthat/setup-global_variable.R
@@ -2,3 +2,4 @@ library(magickGUI)
 options("unit_test_magickGUI" = TRUE)
 img <- wizard
 img2 <- rose
+scale1 <- "500x"