+++ tests/testthat/teardown-global_variable.R
@@ -1,2 +1,4 @@
 options("unit_test_magickGUI" = NULL)
 rm(img)
+rm(img2)
+rm(scale1)