+++ tests/testthat/test_interactive_canny.R
@@ -4,4 +4,5 @@ test_that("interactive_canny",
   expected <- image_canny(img, iniv)
   expect_equal(expected, interactive_canny(img))
   expect_equal(iniv, interactive_canny(img, return_param = TRUE))
+  expect_equal(expected, interactive_canny(img, scale = scale1))
 })