+++ tests/testthat/test_interactive_charcoal.R
@@ -5,4 +5,5 @@ test_that("interactive_charcoal",
   expected <- image_charcoal(img, iniv[1], iniv[2])
   expect_equal(expected, interactive_charcoal(img))
   expect_equal(iniv, interactive_charcoal(img, return_param = TRUE))
+  expect_equal(expected, interactive_charcoal(img, scale = scale1))
 })