+++ tests/testthat/test_interactive_implode.R
@@ -5,4 +5,5 @@ test_that("interactive_implode",
   expected <- image_implode(img, iniv)
   expect_equal(expected, interactive_implode(img))
   expect_equal(iniv, interactive_implode(img, return_param = TRUE))
+  expect_equal(expected, interactive_implode(img, scale = scale1))
 })