+++ tests/testthat/test_interactive_quantize.R
@@ -5,4 +5,5 @@ test_that("interactive_quantize",
   expected <- image_quantize(img, iniv)
   expect_equal(expected, interactive_quantize(img))
   expect_equal(iniv, interactive_quantize(img, return_param = TRUE))
+  expect_equal(expected, interactive_quantize(img, scale = scale1))
 })