+++ tests/testthat/test_interactive_threshold.R
@@ -5,4 +5,5 @@ test_that("interactive_threshold",
   expected <- image_threshold(img, threshold = iniv)
   expect_equal(expected, interactive_threshold(img))
   expect_equal(iniv, interactive_threshold(img, return_param = TRUE))
+  expect_equal(expected, interactive_threshold(img, scale = scale1))
 })