+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * deal with the error of tkimage.create function of tcltk package
 
+* add new parameter scale
+
 ## magickGUI 1.2.0
 
 ### New Functions